Exclude files from templating based on .gitignore
We used to exclude a hardcoded list of "filenames" from templating, so
we didn't chug on trying to rewrite all of `.git` or `node_modules`.
This easily fell out of sync and was couldn't represent complex paths.

Instead, we use `.gitignore`s as a basis for which files to exclude.