Fix `build-package` naming under help
This was hacked around because the command makes sense in kebab-case,
but we use camelCase for modules and exports. Previously `buildPackage`
was displayed as the canonical command name when you ran `skuba help`,
but it wouldn't actually work. This makes `build-package` the actual
command and introduces a simple kebab-case -> camelCase conversion that
could be used for future commands.