Make package prompts unskippable (#76)
This introduces a `noSkip` property that templates can provide to
suppress the option to defer templating until later. This is important
for the package templates because they require a valid `moduleName` to
be substituted into `package.json#/name` or else npm/yarn will crash and
complain about an invalid package name.