Move skuba diving to appropriate step (#79)
This migrates the last of the dependency logic out of the "config" step
into the earlier "dependency" step, so the dependency diff should be
fully accurate now.

This only implements a simple switch of "package means no dive,
application means dive". We can build on this to allow applications to
opt-out, as described in #26.