Avoid stripping of `_` filename prefixes (#119)
This filename manipulation has haunted me as `copyFiles` gets used in
more places. #106 tightened up the matching patterns but a target
repository _could_ have some legitimate reason for storing a filename
with a prefix of `_.` that would still get chomped.

Instead, only trigger this logic on `skuba init` of a built-in template.
There's no reason to do this on `skuba configure`, and external
GitHub-hosted templates shouldn't run into the packaging troubles that
necessitate a `_package.json`.