Document local Lambda deployment (#148)
This was raised by @BryceKK as we don't document how to do this.

To keep it a simple single-variable matter, we now default the `VERSION`
environment variable to `local` if it is not supplied. I generally don't
like defaults but this one is mostly harmless, and we do a similar thing
with `BUILDKITE_COMMIT`.

I took the opportunity to make our READMEs slightly less bad, but they
are still pretty bad.