Sequence Validate action after Release (#169)
This uses GitHub's new `workflow_run` event to sequence our Validate
action to run after the Release action on the default branch. We would
previously run into race conditions when we released a new version, as
Validate would try to run against a version that was in the process of
being released.