Rewrite `dist => lib` in `serverless.yml`s
skuba's default configuration compiles to `lib`, though it's common to
see `dist` and `lib` out in the wild. `skuba configure` already tries to
rewire Dockerfiles to point to this output directory, and now we do the
same for Serverless files.

This implements globbing in our configure modules so we can match on
multiple `serverless*.yml`s in the one repo.