Handcraft a Renovate config
skuba has pretty intense requirements for renovation; most notably, it
bundles templates that need to be tagged and upgraded using different
rules. We've outlasted the usefulness of inherited defaults and our
frankenstein config stopped working entirely a month ago, per #426.

This tries to vastly simplify a copy of the old base config.