Run ESLint and Prettier formatting in process
This took an eternity because I went down a rabbit hole of parallelising
Prettier and had little to show for it based on very rudimentary
benchmarking of skuba itself; it seemed the overhead of spinning up
multiple worker processes was significant. I may give this another shot
with worker threads in future.

While this is undeniably more code that using the ESLint and Prettier
CLIs out of the box, it presents a few opportunities:

- Customising logging output, particularly for `--debug`ging.
- Taking the above point further: supporting Buildkite annotations,
  particularly when `lint`ing.
- A marginally faster execution and smaller resource footprint.
- Likelier compatibility with alternative module resolution approaches
  like Yarn PnP, which can make it harder to reliably locate other
  `.bin` tools when `exec`ing.