Run REPL in process (#534)
When skuba was first conceived, we were unsure if it would have much
staying power. As such, it stuck to extending regular config files and
executing command line tooling; the idea was that consumers could easily
offboard themselves by simply rolling their own underlying commands, and
we could fairly easily sunset skuba if it proved less than useful.

If we fast forward to the present, we are now held back by some of these
initial constraints. For example, sticking with a vanilla `tsc` means
that we can't perform compile-time import alias resolution, and the
abundance of `exec`ing is fairly wasteful as each one spins up its own
Node.js process with a separate V8 instance and all.

This is a simple first step: replace a trivial `exec` with a
pre-existing Node.js API that does the same thing. There's much more to
come in subsequent PRs.