build-package, lint: Add `--serial` flag (#556)
Previously, we had a Buildkite-specific hack to propagate the BUILDKITE
environment variable in order to run these commands serially. This is
potentially surprising behaviour, particularly if the consumer is
propagating the variable for some other purpose.

That point will become less hypothetical once skuba adds built-in
support for Buildkite annotations. We replace the inference with a very
explicit flag.