Reorganise test scripts (#555)
Our integration tests mess with some `process` globals so they need to
be `--runInBand`. This is a bit annoying for the rest of the tests
though, so we keep an integration-less `yarn test` that allows for
parallelism.

This also fixes a loop that can occur when `--watch`ing the integration
tests due to the resulting file I/O in the `integration` directory.