Fix local Changesets formatting weirdness (#677)
We specify a custom Prettier config for our `.changeset` directory which
accepts the double-quoted Markdown files that the GitHub Changesets bot
generates. Unfortunately the Changesets CLI seems to ignore this file
and only uses the root config, so it often generates non-compliant
single-quoted files.

To fix this, we just have Prettier clean up after Changesets.