Register `tsconfig-paths` for development
Note that this does not modify our typical `skuba-dive/register` runtime
hook. The drift between the two may be a bit puzzling in the time being,
but it's somewhat rare for production applications to include
`tsconfig.json` in their final bundle or container. Alignment should
eventually come as we investigate shifting over to `esbuild`, which
supports compile-time `tsconfig-paths` resolution.