configure, init: Remove `yarn-deduplicate`
I don't think we really need this complexity and we'll soon be moving to
a multi-package manager state where it will be more annoying to maintain
customisations like this.

You can configure Renovate to run this command automatically so it's not
a big deal. Not sure if a changeset is warranted even.