Patch `unhandledRejections` listener
Related: #1969, seek-oss/eslint-config-seek#227

Closes: #1356

This makes the assumption that we don't care about doing this for AWS
Lambda functions as the failure mode is less severe for our typical
worker patterns. Is this fair?

Note that if we want to cover Lambda functions, we need to remove the
built-in listener that throws an error:

https://docs.sentry.io/platforms/javascript/guides/aws-lambda/configuration/integrations/unhandledrejection/#on-aws-lambda